server/requirements.txt
PyYAML==6.0.1
Flask==2.3.2
Flask-Cors==3.0.10
waitress==2.1.2
lxml==4.9.1